refactor: moving files for convenience
