+++ store/src/components/modalError/style.css
